+++ {{ cookiecutter.project_name }}/project/settings/production.py
@@ -11,7 +11,8 @@ class Production(ComposedProjectSettings):
     ENVIRONMENT = values.Value("Production", environ_name="DEPLOYED_ENVIRONMENT")
 
     @classmethod
-    def post_setup(cls):
-        super().post_setup()
+    def pre_setup(cls):
         # Expected path for optional Dotenv file
         cls.DOTENV = cls.BASE_DIR / ".env"
+
+        super(Production, cls).pre_setup()